readme edit

+++ src/models/shared.rs
@@ -1,10 +1,11 @@
 use fraction::Decimal;
 
 // Aiming to use the type with the most precision available
-// and consider rounding as a presentation concern.
+// and considering rounding as a presentation concern.
 pub type Amount = Decimal;
 
-pub type TransactionID = u64;
+// As specified by `Rust Test.pdf``
+pub type TransactionID = u32;
 
-// As specified by Rust Test.pdf
+// As specified by `Rust Test.pdf``
 pub type ClientID = u16;
